requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor-5==0.2.16
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-mptt==0.16.0
frozenlist==1.4.1
idna==3.10
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
